Added input validation and error display, converted all buttons to circular Metro style, added french localization

+++ InventoryManager/ViewModels/EditProductViewModel.cs
@@ -1,22 +1,31 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.ComponentModel.Composition;
+using System.Diagnostics;
 using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Xml.Serialization;
 using Caliburn.Micro;
+using InventoryManager.Properties;
 
 namespace InventoryManager.ViewModels
 {
-    [Export(typeof(EditProductViewModel))]
-    public class EditProductViewModel: PropertyChangedBase , IHandle<Dictionary<string, Product>>
+    [Export(typeof (EditProductViewModel))]
+    public class EditProductViewModel : PropertyChangedBase, IHandle<Dictionary<string, Product>>, IDataErrorInfo
     {
+        private const string EditProductMessage = "editProductMessage";
         private readonly IEventAggregator _eventAggregator;
-        private string _editProductMessage = "editProductMessage";
+        private readonly DirectoryInfo _products = new DirectoryInfo(Environment.CurrentDirectory + "\\products");
+        private readonly XmlSerializer _xmlSerializer = new XmlSerializer(typeof (Product));
         private string _oldProductName;
-        private readonly XmlSerializer _xmlSerializer = new XmlSerializer(typeof(Product));
-        readonly DirectoryInfo _products = new DirectoryInfo(Environment.CurrentDirectory + "\\products");
+
+        private string _productCost;
+
+        private string _productName;
+
+        private string _productPrice;
 
         [ImportingConstructor]
         public EditProductViewModel(IEventAggregator eventAggregator)
@@ -25,7 +34,6 @@ namespace InventoryManager.ViewModels
             _eventAggregator.Subscribe(this);
         }
 
-        private string _productName;
         public string ProductName
         {
             get { return _productName; }
@@ -33,10 +41,10 @@ namespace InventoryManager.ViewModels
             {
                 _productName = value;
                 NotifyOfPropertyChange(() => ProductName);
+                NotifyOfPropertyChange(() => CanSaveEditedProduct);
             }
         }
 
-        private string _productPrice;
         public string ProductPrice
         {
             get { return _productPrice; }
@@ -44,10 +52,10 @@ namespace InventoryManager.ViewModels
             {
                 _productPrice = value;
                 NotifyOfPropertyChange(() => ProductPrice);
+                NotifyOfPropertyChange(() => CanSaveEditedProduct);
             }
         }
 
-        private string _productCost;
         public string ProductCost
         {
             get { return _productCost; }
@@ -55,35 +63,45 @@ namespace InventoryManager.ViewModels
             {
                 _productCost = value;
                 NotifyOfPropertyChange(() => ProductCost);
+                NotifyOfPropertyChange(() => CanSaveEditedProduct);
             }
         }
-
+      
+        public void Handle(Dictionary<string, Product> message)
+        {
+            if (!message.ContainsKey(EditProductMessage)) return;
+            _oldProductName = message[EditProductMessage].Name;
+            ProductName = message[EditProductMessage].Name;
+            ProductCost = message[EditProductMessage].Cost.ToString(CultureInfo.InvariantCulture);
+            ProductPrice = message[EditProductMessage].Price.ToString(CultureInfo.InvariantCulture);
+        }
+        public bool CanSaveEditedProduct
+        {
+            get
+            {
+                return !string.IsNullOrEmpty(_productName) && !string.IsNullOrEmpty(_productCost) && !string.IsNullOrEmpty(_productPrice);
+            }
+        } 
         public void SaveEditedProduct()
         {
-            Product product = new Product()
+            var product = new Product
             {
                 Name = ProductName,
                 Cost = Convert.ToDouble(ProductCost),
                 Price = Convert.ToDouble(ProductPrice)
             };
 
-            string pathToDelete = _products.FullName + String.Format("\\{0}.xml", RemoveWhitespace(_oldProductName));
-            File.Delete(pathToDelete);
+            var pathToDelete = _products.FullName + string.Format("\\{0}.xml", RemoveWhitespace(_oldProductName));
+            if (File.Exists(pathToDelete))
+            {
+                File.Delete(pathToDelete);
+            }
 
-            string path = _products.FullName + string.Format("\\{0}.xml", RemoveWhitespace(ProductName));
-            FileStream file = File.Create(path);
+            var path = _products.FullName + string.Format("\\{0}.xml", RemoveWhitespace(ProductName));
+            var file = File.Create(path);
             _xmlSerializer.Serialize(file, product);
             file.Close();
-            _eventAggregator.PublishOnUIThread(new Dictionary<string, Product> { { "saveProductMessage", product } });
-        }
-
-        public void Handle(Dictionary<string, Product> message)
-        {
-            if (!message.ContainsKey(_editProductMessage)) return;
-            _oldProductName = message[_editProductMessage].Name;
-            ProductName = message[_editProductMessage].Name;
-            ProductCost = message[_editProductMessage].Cost.ToString(CultureInfo.InvariantCulture);
-            ProductPrice = message[_editProductMessage].Price.ToString(CultureInfo.InvariantCulture);
+            _eventAggregator.PublishOnUIThread(new Dictionary<string, Product> {{"saveProductMessage", product}});
         }
 
         private string RemoveWhitespace(string inputString)
@@ -94,5 +112,37 @@ namespace InventoryManager.ViewModels
                     .Select(c => c.ToString())
                     .Aggregate((a, b) => a + b);
         }
+
+        public string this[string columnName]
+        {
+            get
+            {
+                string result = null;
+                if(columnName.Equals("ProductName"))
+                {
+                    if (string.IsNullOrEmpty(ProductName))
+                        result = Resources.EnterProductName;
+                    if (!(ProductName is string))
+                        result = Resources.InvalidInput;
+                }
+                if (columnName.Equals("ProductCost"))
+                {
+                    if (string.IsNullOrEmpty(ProductCost))
+                        result = Resources.EnterProductCost;
+                    if (ProductCost.All(c => !(char.IsDigit(c) || c.Equals('.'))))
+                        result = Resources.InvalidInput;
+                }
+                if (columnName.Equals("ProductPrice"))
+                {
+                    if (string.IsNullOrEmpty(ProductPrice))
+                        result = Resources.EnterProductPrice;
+                    if (ProductPrice.All(c => !(char.IsDigit(c) || c.Equals('.'))))
+                        result = Resources.InvalidInput;
+                }
+                return result;
+            }
+        }
+
+        public string Error { get { return string.Empty; } }
     }
 }
